Move oid_cache one directory up
