+++ src/pg_diffix.c
@@ -4,8 +4,8 @@
 #include "pg_diffix/auth.h"
 #include "pg_diffix/config.h"
 #include "pg_diffix/hooks.h"
+#include "pg_diffix/oid_cache.h"
 #include "pg_diffix/utils.h"
-#include "pg_diffix/query/oid_cache.h"
 
 #include <limits.h>
 #if __WORDSIZE != 64
